Abstract SVG bytes & dims to extension
